# openapi/diasporaland-v1.yaml
openapi: 3.1.0
info:
  title: Landintel Property Intelligence API
  version: 1.0.0
  description: |
    Official canonical REST API for Landintel / DiasporaLand:
    Global Property Due-Diligence, Document Intelligence, Currency & Localization Engine.
servers:
  - url: https://temperature-feels-republic-symphony.trycloudflare.com/api/v1
    description: Live Development & Staging Server
  - url: https://landintel.ng/api/v1
    description: Production Edge Server
paths:
...
  /auth/me:
    get:
      summary: Get Current Authenticated User Context
      operationId: getCurrentUser
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /cases:
    post:
      summary: Create Property Case
      operationId: createCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: Property case created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    get:
      summary: List Property Cases
      operationId: listCases
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 25
      responses:
        '200':
          description: Paginated collection of property cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCollectionResponse'
...
  /exchange-rates:
    get:
      summary: Get Live Exchange Rates
      operationId: getExchangeRates
      parameters:
        - name: base
          in: query
          schema:
            type: string
            default: USD
        - name: quote
          in: query
          schema:
            type: string
            default: NGN
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRateResponse'

  # Languages & Localizations
...
  /support/requests:
    post:
      summary: Submit Support Inquiry or Report Issue
      operationId: createSupportRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupportRequest'
      responses:
        '201':
          description: Ticket registered and admin notified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportRequestResponse'

    get:
      summary: List Support Requests
      operationId: listSupportRequests
      responses:
        '200':
          description: List of support requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SupportRequest'
                  meta:
                    $ref: '#/components/schemas/Meta'

  # Customer Notifications
...
components:
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required:
        - requestId
...
    CurrentUserResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/AuthUser'
            session:
              $ref: '#/components/schemas/Session'
        meta:
          $ref: '#/components/schemas/Meta'
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        emailVerified:
          type: boolean
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
    CreateCaseRequest:
      type: object
      required:
        - name
        - countryCode
        - propertyType
      properties:
        name:
          type: string
          example: Lekki Phase 1 Residential Parcel
        countryCode:
          type: string
          example: NG
        jurisdiction:
          type: object
          properties:
            regionCode:
              type: string
              example: LA
            districtCode:
              type: string
              example: ETI-OSA
            city:
              type: string
              example: Lekki
        propertyType:
          type: string
          enum: [LAND, RESIDENTIAL, COMMERCIAL, AGRICULTURAL, MIXED_USE, OTHER]
          example: LAND
    CaseResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Case'
        meta:
          $ref: '#/components/schemas/Meta'
    CaseCollectionResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Case'
        meta:
          $ref: '#/components/schemas/Meta'
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
          example: DL-2026-000123
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, UPLOADING, PROCESSING, ANALYZING, PARTIALLY_ANALYZED, ANALYZED, REQUIRES_REVIEW, REPORT_READY, ARCHIVED, FAILED]
        countryCode:
          type: string
        propertyType:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ExchangeRateResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            baseCurrency:
              type: string
            quoteCurrency:
              type: string
            rate:
              type: string
            source:
              type: string
...
    CreateSupportRequest:
      type: object
      required:
        - email
        - subject
        - message
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        category:
          type: string
          enum: [GENERAL_SUPPORT, PROPERTY_ANALYSIS, VERIFICATION, DOCUMENT, REPORT, PAYMENT, BILLING, OTHER]
        subject:
          type: string
        message:
          type: string
        caseId:
          type: string
          format: uuid
          nullable: true
    SupportRequestResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SupportRequest'
        meta:
          $ref: '#/components/schemas/Meta'
    SupportRequest:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
          example: DL-SUP-2026-000184
        name:
          type: string
        email:
          type: string
        category:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        createdAt:
          type: string
          format: date-time
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input parameters provided.
            details:
              type: object
            requestId:
              type: string
              format: uuid
          required:
            - code
            - message
            - requestId
  responses:
    ValidationError:
      description: Bad request validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...